- Moved Formatter feature into Launcher feature
