--- src/main/java/com/neopragma/cobolcheck/features/Formatter/DataTransferObjects/DataTransferObjectStyle.java
@@ -1,6 +1,0 @@
-package com.neopragma.cobolcheck.features.Formatter.DataTransferObjects;
-
-public enum DataTransferObjectStyle {
-    directOutput,
-    JUnit
-}
