+++ src/main/java/com/neopragma/cobolcheck/features/launcher/Formatter/DataTransferObjects/DataTransferObjectStyle.java
@@ -1,0 +1,6 @@
+package com.neopragma.cobolcheck.features.launcher.Formatter.DataTransferObjects;
+
+public enum DataTransferObjectStyle {
+    directOutput,
+    JUnit
+}
