+++ src/main/java/com/neopragma/cobolcheck/features/launcher/Formatter/Formats/TestOutputFormat.java
@@ -1,4 +1,4 @@
-package com.neopragma.cobolcheck.features.Formatter.Formats;
+package com.neopragma.cobolcheck.features.launcher.Formatter.Formats;
 
 /* Formats that test results can be given in
  * Every enum value is a valid file extension, as it is used to
